requirements.txt
Django>=5.1.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter>=25.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9 